protocols/unix: Implement MSG_NOSIGNAL.
Same logic as in the TCP protocol module.

This fixes a regression from hrev57383: curl now works
again. (Perhaps it shouldn't be called a "regression",
though, but instead an uncovered bug / missing feature.)

Fixes #18666.

+++ src/add-ons/kernel/network/protocols/unix/UnixDatagramEndpoint.cpp
@@ -215,7 +215,7 @@ UnixDatagramEndpoint::Send(const iovec* vecs, size_t vecCount,
 	TRACE("[%" B_PRId32 "] %p->UnixDatagramEndpoint::Send()\n",
 		find_thread(NULL), this);
 
-	if ((flags & ~(MSG_DONTWAIT)) != 0)
+	if ((flags & ~(MSG_DONTWAIT | MSG_NOSIGNAL)) != 0)
 		return EOPNOTSUPP;
 
 	bigtime_t timeout = 0;
@@ -307,7 +307,7 @@ UnixDatagramEndpoint::Send(const iovec* vecs, size_t vecCount,
 
 	switch (result) {
 		case EPIPE:
-			if (gStackModule->is_syscall())
+			if (gStackModule->is_syscall() && (flags & MSG_NOSIGNAL) == 0)
 				send_signal(find_thread(NULL), SIGPIPE);
 			break;
 		case B_TIMED_OUT:
